requirements.txt
discord.py
bardapi
python-dotenv